Make the LED timer bit width a compile time option.
Avoid an unnecessary cast.

+++ src/usb.rs
@@ -1,8 +1,9 @@
 //! USB for GPS ref.
 //! Endpoints:
 //! 0 : Control, as always.
-//! 01, 81: CDC ACM data transfer, bulk
+//! 01, 81: CDC ACM data transfer for GPS serial, bulk
 //! 82: CDC ACM interrupt IN (to host).
+//! 03, 83: Device commands.
 
 use crate::cpu::interrupt;
 
@@ -93,11 +94,11 @@ const INTR_TX_OFFSET: usize = 0x40;
 const MAIN_RX_OFFSET: usize = 0x200;
 const MAIN_TX_OFFSET: usize = 0x240;
 
-const BULK_RX_BUF: *mut u8 = (USB_SRAM_BASE + BULK_RX_OFFSET) as *mut u8;
-const BULK_TX_BUF: *mut u8 = (USB_SRAM_BASE + BULK_TX_OFFSET) as *mut u8;
-const INTR_TX_BUF: *mut u8 = (USB_SRAM_BASE + INTR_TX_OFFSET) as *mut u8;
-const MAIN_RX_BUF: *mut u8 = (USB_SRAM_BASE + MAIN_RX_OFFSET) as *mut u8;
-const MAIN_TX_BUF: *mut u8 = (USB_SRAM_BASE + MAIN_TX_OFFSET) as *mut u8;
+const BULK_RX_BUF: *mut u8  = (USB_SRAM_BASE + BULK_RX_OFFSET) as *mut u8;
+const BULK_TX_BUF: *mut u32 = (USB_SRAM_BASE + BULK_TX_OFFSET) as *mut u32;
+const INTR_TX_BUF: *mut u8  = (USB_SRAM_BASE + INTR_TX_OFFSET) as *mut u8;
+const MAIN_RX_BUF: *mut u8  = (USB_SRAM_BASE + MAIN_RX_OFFSET) as *mut u8;
+const MAIN_TX_BUF: *mut u8  = (USB_SRAM_BASE + MAIN_TX_OFFSET) as *mut u8;
 
 fn chep_ser () -> &'static stm32h503::usb::CHEPR {chep_ref(1)}
 fn chep_intr() -> &'static stm32h503::usb::CHEPR {chep_ref(2)}
